Split main.go
